fix(add-credit-card-modal): move to organization domain
