+++ libs/shared/console-shared/src/index.ts
@@ -8,7 +8,6 @@ export * from './lib/entrypoint-cmd-inputs/ui/entrypoint-cmd-inputs'
 export * from './lib/flow-create-port/ui/flow-create-port/flow-create-port'
 export * from './lib/flow-create-port/ui/crud-modal/crud-modal'
 export * from './lib/flow-create-port/utils/port-healthcheck'
-export * from './lib/add-credit-card-modal/feature/add-credit-card-modal-feature'
 export * from './lib/layout-logs/layout-logs'
 export * from './lib/layout-logs/card-cluster/card-cluster'
 export * from './lib/database-settings-resources/ui/database-settings-resources/database-settings-resources'
